Release 6 (#89)
* update: version
* fix(android/gradlew): Permission denied
* update: readme
* improve: audio decoding (#75)
* improve(audio): speed/tempo filter (#76)
* improve(audio): autodetect device connections (#77)
* update: readme
* fix: sometimes crashes when toggling media tracks (#78)
* add(api): public endpoints to play/pause (#79)
* improve(LVP_GetMediaDetails): add option to skip generating a thumbnail (#80)
* add(api): endpoint to get media meta (#81)
* improve(audio): autodetect device connections (#77)
* improve(seeking): external subtitle files (#82)
* add(cmake): options to disable test, av1 and libass (#83)
* fix(subs): overlapping hdmv/pgs timestamps (#84)
* fix(subs): crashes on disable (#85)
* update: version
* update(doc): min compiler version for cpp 20
* fix(build): library avcodec not found (#86)
* update(doc): readme
* fix(android): sdl version mismatch (#88)
* fix(build): library freetype not found (#87)
* update(build): license file name

+++ src/LVP_TimeOut.h
@@ -12,12 +12,12 @@ namespace LibVoyaPlayer
 		class LVP_TimeOut
 		{
 		public:
-			LVP_TimeOut(uint32_t timeOut = 5000U);
+			LVP_TimeOut(uint64_t timeOut = 5000ULL);
 
 		private:
 			bool     started;
-			uint32_t startTime;
-			uint32_t timeOut;
+			uint64_t startTime;
+			uint64_t timeOut;
 
 		public:
 			static int InterruptCallback(void* data);
